+++ src/_middleware/session.derive.ts
@@ -2,18 +2,19 @@
 import { verifyRequestOrigin } from "oslo/request";
 import { lucia } from "~config/lucia";
 
-export const sessionDerive = async ({ request:{ headers, method }, cookie, elysia_jwt }:any) => {
+export const sessionDerive = async ({ request:{ headers, method }, cookie, jwt, locals }:any) => {
     // console.debug("Method: ", method);
     // console.debug("Headers: ", headers);
     
     // Check if the Authentication-Method header is present
-    const authMethod = headers.get('Authentication-Method');
+    // const authMethod = headers.get('Authentication-Method');
+    const authMethod = headers.get('X-Client-Type');
     if (!authMethod) {
         // If the header is missing, return an error response
         return { user: null, session: null, authMethod };
     }
 
-    if (authMethod === 'Cookie') {
+    if (authMethod === "Cookie") {
         // CSRF check
 		if (method !== "GET") {
 			const originHeader = headers.get("Origin"); // e.g: http://localhost:3000
@@ -23,18 +24,20 @@ export const sessionDerive = async ({ request:{ headers, method }, cookie, elysi
 			if (!originHeader || !hostHeader || !verifyRequestOrigin(originHeader, [hostHeader])) {
                 return {
 					user: null,
-					session: null, authMethod
+					session: null,
+                    authMethod: authMethod
 				};
 			}
 		}
 
         // use headers instead of Cookie API to prevent type coercion
-        const cookieHeader = headers.get("Cookie");
-        const sessionId = lucia.readSessionCookie(cookieHeader ?? "");
+        const cookieHeader = headers.get("Cookie") ?? "";
+        const sessionId = lucia.readSessionCookie(cookieHeader);
         if (!sessionId) {
             return {
                 user: null,
-                session: null, authMethod
+                session: null,
+                authMethod: authMethod
             };
         }
 
@@ -70,12 +73,24 @@ export const sessionDerive = async ({ request:{ headers, method }, cookie, elysi
             return { user: null, session: null, authMethod };
         }
     
-        const tokenUser = await elysia_jwt.verify(token);
+        const tokenUser = await jwt.verify(token);
         
         if(!tokenUser){
             return { user: null, session: null, authMethod };
         }
         
         return { user:tokenUser, session:null, authMethod };
-    }   
+    } else {
+        // Fallback to User-Agent parsing
+        const userAgent = headers.get('user-agent');
+        if (/mobile|android|ios|flutter/i.test(userAgent)) {
+            locals.isNativeApp = true;
+        } else if (/mozilla|chrome|safari|firefox|edge/i.test(userAgent)) {
+            locals.isBrowser = true;
+        } else {
+            locals.isUnknown = true;
+            // If an unsupported authentication method is specified, return an error response
+            return { user: null, session: null, authMethod };
+        }
+    }
 }