+++ src/_modules/auth/auth.controller.ts
@@ -12,16 +12,18 @@ import { GitHubUserResult, GoogleUserResult, OAuth2Providers } from "./auth.mode
 import { generateCodeVerifier, generateState, OAuth2RequestError } from "arctic";
 import { parseCookies, serializeCookie } from "oslo/cookie";
 import { splitWords, usernameFromEmail } from "~utils/utilities";
-import { blacklistToken, cache, redisSet } from "~config/redis";
+import { blacklistToken, redisMessagingService, redisSet } from "~config/redis";
 import { UsersService } from "~modules/users";
+import { AuthenticationError, ConflictError, NotFoundError } from "src/_exceptions/custom_errors";
 
 export class AuthController {
     // private authService: AuthService;
     private userService: UsersService;
+    private authService = AuthService.getInstance();
     url = `${Bun.env.NODE_ENV === 'production' ? 'https' : 'http'}://${Bun.env.HOST ?? '127.0.0.1'}:${Bun.env.PORT ?? 3000}${consts.api.versionPrefix}${consts.api.version}`;
 
-    constructor(private authService: AuthService) {
-        // this.authService = authService;
+    constructor() {
+        this.authService = AuthService.getInstance();
         this.userService = new UsersService();
     }
 
@@ -42,7 +44,7 @@ export class AuthController {
         return isBrowser ? Bun.file('public/register.html') : { message: 'Use POST instead'}
     }
 
-    login = async ({ set, request:{headers}, body, elysia_jwt, authMethod, cookie:{lucia_auth_cookie}, session }: any) => {
+    login = async ({ set, request:{headers}, body, jwt, authMethod, cookie:{lucia_auth_cookie}, session }: any) => {
         const {email, password, rememberme} = body;
         
         try {
@@ -53,21 +55,24 @@ export class AuthController {
             // find user by Key, and validate password
             let userExists:any = await db.user.findUnique({ where: { email: email.toLowerCase() }, include: { profile: true } });
             if(!userExists){
-                set.status = HttpStatusEnum.HTTP_404_NOT_FOUND;
-                return { message: 'Invalid User credentials' };
+                throw new NotFoundError("Invalid User credentials");
+                // set.status = HttpStatusEnum.HTTP_404_NOT_FOUND;
+                // return { message: 'Invalid User credentials' };
             }
 
             const isMatch = await Bun.password.verify(password, userExists.hashedPassword);
 
             if(!isMatch){
-                set.status = HttpStatusEnum.HTTP_404_NOT_FOUND;
-                return { message: 'Invalid credentials' };
+                throw new NotFoundError("Invalid credentials");
+                // set.status = HttpStatusEnum.HTTP_404_NOT_FOUND;
+                // return { message: 'Invalid credentials' };
             }
 
 
             if(userExists.isActive === false || !userExists.isActive){
-                set.status = HttpStatusEnum.HTTP_403_FORBIDDEN;
-                return { message: 'User access is revoked.', data: userExists.isComment ?? "N/A" };
+                throw new AuthenticationError("User access is revoked", 403, userExists.isComment ?? "N/A");
+                // set.status = HttpStatusEnum.HTTP_403_FORBIDDEN;
+                // return { message: 'User access is revoked.', note: userExists.isComment ?? "N/A" };
             }
 
             const sessions = await lucia.getUserSessions(userExists.id);
@@ -93,21 +98,40 @@ export class AuthController {
                 sessions: sessions.length
             }
 
-            const tokenOrCookie = await this.authService.createDynamicSession(authMethod, elysia_jwt, userExists, headers, rememberme);
+            // Generate access token (JWT) using logged-in user's details
+            const accessToken = await jwt.sign({
+                id: userExists.id,
+                firstname: userExists.firstname,
+                lastname: userExists.lastname,
+                roles: userExists.roles,
+                emailVerified: userExists.emailVerified,
+                createdAt: userExists.createdAt,
+                profileId: userExists.profile?.id ?? null
+            });
+
+            
+
+            const {id} = await this.authService.createLuciaSession(userExists.id, headers, rememberme);
+            const sessionCookie = lucia.createSessionCookie(id);
+
+            redisMessagingService.publish('user-events', {
+                action: "user_logged-in",
+                user: sanitizedUser
+            });
+
+            // RAW redis function
+            const rr = await redisSet(`user:${userExists.id}`, sanitizedUser);
+
+
+            const tokenOrCookie = await this.authService.createDynamicSession(authMethod, jwt, userExists, headers, rememberme);
             if(authMethod === 'JWT'){
                 set.headers["Authorization"] = `Bearer ${tokenOrCookie}`;
             } else if (authMethod === 'Cookie'){
                 set.headers["Set-Cookie"] = tokenOrCookie.serialize();
             }
 
-            // DELETE PASSWORD FROM OBJECT FOR SAFETY
-            delete userExists.hashedPassword;
-            
-            // RAW redis function
-            const rr = await redisSet(`user:${userExists.id}`, userExists);
-            
             set.status = HttpStatusEnum.HTTP_200_OK;
-            return { data: userExists, message: 'Successfully logged in', note: { sessions: sessions.length} };
+            return { data: sanitizedUser, message: 'Successfully logged in', note: { sessions: sessions.length} };
         } catch (e:any) {
             console.error(e);
             
@@ -132,13 +156,14 @@ export class AuthController {
             }
 
             console.error(e);
-            set.status = HttpStatusEnum.HTTP_500_INTERNAL_SERVER_ERROR;
-            return { message: "An unknown login error occurred" };
+            throw e;
+            // set.status = HttpStatusEnum.HTTP_500_INTERNAL_SERVER_ERROR;
+            // return { message: "An unknown login error occurred" };
         }
     }
 
 
-    signup = async({ set, body } :any) => {
+    signup = async({ set, body, request:{ headers } } :any) => {
         const { firstname, lastname, email, phone, password, confirmPassword} = body;
         
         // Create User 
@@ -161,7 +186,7 @@ export class AuthController {
             });
 
             const uid = generateId(16);
-            const newUser: any = await db.user.create({ data: {
+            const newUser: User|null = await db.user.create({ data: {
                 id: uid,
                 firstname,
                 lastname,
@@ -170,7 +195,7 @@ export class AuthController {
                 phone: (autoUser?.phone ?? phone) ?? null,
                 emailVerified: false,
                 isActive: true,
-                roles: autoUser?.roles,
+                roles: autoUser ? autoUser?.roles : [Role.GUEST],
                 hashedPassword
             } }) as User;
 
@@ -187,23 +212,28 @@ export class AuthController {
                     });
             }
 
-            // if(autoUser?.supportLevel && autoUser?.supportLevel < 1){
-            //     await db.autoEnrol.update({ where: { email: autoUser?.email}, data: { isActive: false, isComment: `Used for User Registration at ${new Date()}` } });
-            // }
+            if(autoUser?.supportLevel && autoUser?.supportLevel < 1){
+                await db.autoEnrol.update({ where: { email: autoUser?.email}, data: { isActive: false, isComment: `Used for User Registration at ${new Date()}` } });
+            }
 
-            // const session = await this.authService.createLuciaSession(result.id, headers);
-            // const sessionCookie = lucia.createSessionCookie(session.id);
+            const session = await this.authService.createLuciaSession(newUser.id, headers);
+            const sessionCookie = lucia.createSessionCookie(session.id);
 
-            // DELETE PASSWORD FROM RETURN OBJECT
-            delete newUser.hashedPassword;
+            const sanitizedUser: Partial<User> = await this.authService.sanitizeUserObject(newUser);
+
+            redisMessagingService.publish('user-events', {
+                action: "user-registered",
+                user: sanitizedUser
+            });
 
 
             set.status = HttpStatusEnum.HTTP_201_CREATED;
             return {
-                data: newUser,
+                data: sanitizedUser,
                 message: autoUser ? `${(autoUser.roles?.length ?? 'Nil')} roles Account successfully created, ${firstname} ${lastname}` : `Guest Account successfully created (${firstname} ${lastname})`,
                 note: 'User account created' + autoUser ? ' via Auto-enrol' : '. No Auto-enrol'
             };
+
         } catch (e) {
             console.error(e);
 
@@ -218,15 +248,16 @@ export class AuthController {
                 e instanceof Error
                 && e.message === "AUTH_DUPLICATE_KEY_ID"
             ) {
-                set.status = HttpStatusEnum.HTTP_409_CONFLICT;
-                return { message: "that email address is already taken" };
+                throw new ConflictError("That email address is taken", HttpStatusEnum.HTTP_409_CONFLICT);
+                // set.status = HttpStatusEnum.HTTP_409_CONFLICT;
+                // return { message: "that email address is already taken" };
             }
-            set.status = 500
-            return { message: "An unknown auth error occurred" };
+            
+            throw e;
         }
     }
 
-    logout = async ({ set, request:{headers}, session, cookie, elysia_jwt, authMethod }: any) => {
+    logout = async ({ set, request:{headers}, session, cookie, jwt, authMethod }: any) => {
 
         try {
             // Determine the authentication method
@@ -251,7 +282,7 @@ export class AuthController {
                 
             } else if (authMethod === 'JWT') {
                 // Handle JWT-based authentication
-                if (!await elysia_jwt.verify(token)) {
+                if (!await jwt.verify(token)) {
                     set.status = HttpStatusEnum.HTTP_401_UNAUTHORIZED;
                     return { message: 'Invalid or expired token' };
                 }
@@ -260,7 +291,7 @@ export class AuthController {
                 await blacklistToken(token, 60 * 60); // Adjust the expiry as needed
     
                 // Optionally sign out the JWT token
-                await elysia_jwt.sign(null);
+                await jwt.sign(null);
     
                 // No need to handle cookies for JWT auth
             } else {