+++ src/_modules/root/app.models.ts
@@ -1,4 +1,5 @@
 export interface CustomError extends Error {
+  status: number;
     validator?: {
       schema?: {
         properties?: any;