+++ src/index.ts
@@ -1,4 +1,4 @@
-import { Elysia, error } from "elysia";
+import { Elysia } from "elysia";
 
 // Configurations
 import consts from "~config/consts";
@@ -10,14 +10,13 @@ import { rateLimit } from "elysia-rate-limit";
 import cron, { Patterns } from "@elysiajs/cron";
 import cors from "@elysiajs/cors";
 import { helmet } from "elysia-helmet";
-import cookie from "@elysiajs/cookie";
 import jwt from "@elysiajs/jwt";
 import { staticPlugin } from '@elysiajs/static';
 import { htmx } from "elysia-htmx";
 
 // Middleware
 import { bootLogger, gracefulShutdown, requestLogger } from "~utils/systemLogger";
-import { ErrorMessages } from "~middleware/errorMessages";
+import { errorMessages } from "~middleware/errorMessages";
 import { checkMaintenanceMode } from "~middleware/lifecycleHandlers";
 import customResponse from "~middleware/customResponse";
 import { sessionDerive } from "~middleware/session.derive";
@@ -29,16 +28,19 @@ import { ip } from "elysia-ip";
 import { Logestic } from "logestic";
 import { FilesController } from "~modules/files";
 import { AuthService } from "./_modules";
-
-
-const authService = AuthService.getInstance();
-const files = new FilesController();
+import { DatabaseError } from "./_exceptions/custom_errors";
+import { redisMessagingService } from "~config/redis";
 
 
 try {
   console.log("Initializing Elysia...");
+
+  const authService = AuthService.getInstance();
+  const files = new FilesController();
+
   if (import.meta.main) {
     const PORT = Bun.env.PORT || 3000;
+
     const app = new Elysia({
       name: consts.server.name,
       prefix: `/v${consts.api.version}`,
@@ -46,6 +48,9 @@ try {
       detail: { description: `${consts.server.name} Server API` }
     })
 
+    // Error Handling
+    .onError(errorMessages)
+    
     // State
     .state('maintenanceMode', Bun.env.MAINTENANCE_MODE === 'true' || false)
     .state('timezone', String(Bun.env.TZ || 'Europe/London'))
@@ -53,15 +58,9 @@ try {
 
     /* Extensions */
 
-    
     // Fancy logs
-    .use(Logestic.preset('fancy'))
+    .use(Logestic.preset("common"))
 
-    // Log errors
-    // .use(logger({ 
-    //   level: 'error',
-    //   // file: "./my.log", // fileLogger
-    // }))
 
     // Swagger
     .use(swagger({ autoDarkMode: true, documentation: {
@@ -78,8 +77,7 @@ try {
       methods: ['OPTIONS', 'GET', 'PUT', 'POST', 'PATCH'],
       credentials: true,
       origin: /localhost.*/,
-      // origin: (ctx) => ctx.headers.get('Origin'),
-      allowedHeaders: ['Content-Type', 'Authorization', 'Credentials', 'Origin', 'Host', 'os', 'ipCountry', 'X-Forwarded-For', 'X-Real-IP', 'X-Custom-Header', 'requestIP', 'Authentication-Method']
+      allowedHeaders: ['Content-Type', 'Authorization', 'Credentials', 'Origin', 'Host', 'os', 'ipCountry', 'X-Forwarded-For', 'X-Real-IP', 'X-Custom-Header', 'requestIP', 'Authentication-Method', 'X-Client-Type' ]
     }))
 
     // Helmet security (might conflict with swagger)
@@ -99,7 +97,7 @@ try {
     // JWT
     .use(
       jwt({
-          name: 'elysia_jwt',
+          name: 'jwt',
           secret: Bun.env.JWSCRT!,
           exp: `${consts.auth.jwtMaxAge}d`
       })
@@ -146,25 +144,39 @@ try {
     // HTMX plugin
     .use(htmx())
 
-    // Get IP of client and add to context
-    .use(ip({ checkHeaders: ["X-Forwarded-For", "X-Real-IP", "requestIP", "Authentication-Method"] }))
 
-    
+
+    // Get IP of client and add to context
+    // .use(ip({ checkHeaders: ["X-Forwarded-For", "X-Real-IP", "requestIP", "Authentication-Method"] }))
+
+
 
     // Life cycles
     .derive(sessionDerive) // Adds User and Session data to context - from token/cookie
     .onBeforeHandle([checkMaintenanceMode]) // Checks if server is in maintenance mode
-    .onError(({ code, error, set }:any) => ErrorMessages(code, error, set)) // General Error catching system
     .mapResponse(customResponse)
     .onStop(gracefulShutdown);
 
-    console.log("Initializing Elysia... Done!");
+
 
     // ROUTES
     registerControllers(app as any);
 
-    process.on('SIGINT', app.stop);
+
+    process.on('SIGINT', () => {
+      console.log('Stopping App...');
+      // close Redis system
+      redisMessagingService.close();
+      process.exit(0);
+      // app.stop();
+    });
+    process.on('SIGSEGV', app.stop);
     process.on('SIGTERM', app.stop);
+    process.on('SIGKILL', (e) => { console.error(e); app.stop; });
+
+
+    console.log("Initializing Elysia... Done!");
+
 
     // initialize server
     app.listen(PORT, bootLogger);